Remove old pricing page components and content
