+++ src/components/PricingPage/SectionOne/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const SectionOne = () => {
+    return <div>SectionOne</div>;
+};
+
+export default SectionOne;
