+++ src/components/PricingPage/index.tsx
@@ -1,0 +1,3 @@
+import SectionOne from './SectionOne';
+
+export { SectionOne };
