+++ src/pages/pricing.tsx
@@ -1,31 +1,35 @@
-import '../layouts/Pricing/style.less';
-
 import React from 'react';
-
-import Header from '../layouts/Pricing/Head';
-import Wrapper from '../layouts/Pricing/Wrapper';
-
-import { PricingCalculator } from '../components/PricingCalculator';
-import CostCalculator from '../layouts/Pricing/Sections/CostCalculator';
-import FrequentlyQuestions from '../layouts/Pricing/Sections/FrequentlyQuestions';
-import Hero from '../layouts/Pricing/Sections/Hero';
-import Plans from '../layouts/Pricing/Sections/Plans';
-import PricingComparison from '../layouts/Pricing/Sections/PricingComparison';
+import { graphql, useStaticQuery } from 'gatsby';
+import Layout from '../components/Layout';
+import Seo from '../components/seo';
+import { SectionOne } from '../components/PricingPage';
 
 const PricingPage = () => {
     return (
-        <Wrapper>
-            <Hero />
-            <Plans />
-            <PricingCalculator.Provider>
-                <CostCalculator />
-            </PricingCalculator.Provider>
-            <PricingComparison />
-            <FrequentlyQuestions />
-        </Wrapper>
+        <Layout>
+            <SectionOne />
+        </Layout>
     );
 };
 
-export const Head = Header;
+export const Head = () => {
+    const { metaImg } = useStaticQuery(graphql`
+        query {
+            metaImg: file(relativePath: { eq: "pricing/graphic_parent.png" }) {
+                childImageSharp {
+                    gatsbyImageData(layout: FIXED, width: 618)
+                }
+            }
+        }
+    `);
+
+    return (
+        <Seo
+            title="Pricing"
+            description="Reduce your data costs and latency with managed streaming CDC and ETL pipelines."
+            image={metaImg.childImageSharp.gatsbyImageData.images.fallback.src}
+        />
+    );
+};
 
 export default PricingPage;
